[refactor] #22 정치인Util 코드 일부 정치인 정보 Service로 이동
